latest_site_package/requirements.txt
fastapi>=0.110,<1.0
uvicorn[standard]>=0.27,<1.0
pydantic-settings>=2.2,<3.0
sqlalchemy>=2.0,<3.0
asyncpg>=0.29,<1.0
aioboto3>=12.4,<13.0
httpx>=0.27,<1.0
pyjwt[crypto]>=2.8,<3.0
openai>=1.12,<2.0
python-multipart>=0.0.6,<0.1
alembic>=1.13,<2.0
gunicorn>=21.2,<22.0
pydantic-core>=2.18,<3.0
